Add support to non local fields
Right now only local fields are being watched, so changes in subclasses
won't be caught if the field changing belongs to the parent class. Just
changing _as_dict implementation this could be supported too.

Also using field.attname in order to get all the values, since this
simplifies the code a little bit.